feat: extract solana unwrap tip component
