+++ apps/web/src/views/SwapSimplify/InfinitySwap/UnwrapTips.tsx
@@ -1,0 +1,73 @@
+import React, { useCallback } from 'react'
+import { CircleInfo, Flex, Text, useToast } from '@pancakeswap/uikit'
+import { useTranslation } from '@pancakeswap/localization'
+import { WSOLMint } from '@pancakeswap/solana-core-sdk'
+import { NonEVMChainId } from '@pancakeswap/chains'
+import { useWallet } from '@solana/wallet-adapter-react'
+import { createCloseAccountInstruction } from '@solana/spl-token-0.4'
+import { Transaction } from '@solana/web3.js'
+import { useSolanaTokenBalance, useRefreshSolanaTokenBalances } from 'state/token/solanaTokenBalances'
+import { useSolanaConnectionWithRpcAtom } from 'hooks/solana/useSolanaConnectionWithRpcAtom'
+import useAccountActiveChain from 'hooks/useAccountActiveChain'
+import { useSwapCurrency } from '../../Swap/V3Swap/hooks/useSwapCurrency'
+
+export const UnwrapTips: React.FC = () => {
+  const { t } = useTranslation()
+  const { publicKey, signTransaction } = useWallet()
+  const connection = useSolanaConnectionWithRpcAtom()
+  const { solanaAccount } = useAccountActiveChain()
+  const { balance: wsolBalance } = useSolanaTokenBalance(solanaAccount, WSOLMint.toBase58())
+  const refreshSolanaBalances = useRefreshSolanaTokenBalances(solanaAccount)
+  const { toastSuccess, toastError } = useToast()
+  const [inputCurrency, outputCurrency] = useSwapCurrency()
+
+  const isSolanaSwap =
+    inputCurrency?.chainId === NonEVMChainId.SOLANA || outputCurrency?.chainId === NonEVMChainId.SOLANA
+  const showUnwrapTip = isSolanaSwap && wsolBalance.gt(0)
+
+  const handleUnwrap = useCallback(async () => {
+    try {
+      if (!publicKey || !signTransaction) throw new Error('Wallet not connected')
+      const accounts = await connection.getTokenAccountsByOwner(publicKey, { mint: WSOLMint })
+      if (accounts.value.length === 0) return
+      const tx = new Transaction()
+      accounts.value.forEach(({ pubkey }) => {
+        tx.add(createCloseAccountInstruction(pubkey, publicKey, publicKey))
+      })
+      tx.feePayer = publicKey
+      const { blockhash } = await connection.getLatestBlockhash()
+      tx.recentBlockhash = blockhash
+      const signed = await signTransaction(tx)
+      const sig = await connection.sendRawTransaction(signed.serialize())
+      await connection.confirmTransaction(sig)
+      toastSuccess(t('Success!'), t('Unwrapped WSOL to SOL'))
+      refreshSolanaBalances()
+    } catch (e: any) {
+      toastError(t('Failed'), e?.message ?? 'Unwrap failed')
+    }
+  }, [publicKey, signTransaction, connection, toastSuccess, toastError, t, refreshSolanaBalances])
+
+  if (!showUnwrapTip) return null
+
+  return (
+    <Flex
+      mb="12px"
+      fontSize="14px"
+      alignItems="center"
+      px="8px"
+      py="8px"
+      backgroundColor="rgba(0,0,0,0.05)"
+      borderRadius="8px"
+    >
+      <CircleInfo mr="4px" />
+      <Text>
+        {t('You have %amount% WSOL that you can ', { amount: wsolBalance.dividedBy(1e9).toFixed(6) })}
+        <Text as="span" color="primary" cursor="pointer" onClick={handleUnwrap}>
+          {t('unwrap')}
+        </Text>
+      </Text>
+    </Flex>
+  )
+}
+
+export default UnwrapTips
